:recycle: refactor: mudando a rota de criar psicologo para registro em auth
