+++ src/psychologists/psychologists.controller.spec.ts
@@ -75,98 +75,7 @@ describe('PsychologistsController', () => {
     expect(service).toBeDefined();
   });
 
-  describe('ValidateFileType', () => {
-    it('Should throw an error if profilePicture is not an image format', () => {
-      const profilePicture = {
-        originalname: 'profile.pdf',
-        buffer: Buffer.from(''),
-      } as Express.Multer.File;
-      expect(() =>
-        controller.validateUploadedFile(profilePicture, ['jpg', 'jpeg', 'png'])
-      ).toThrow(BadRequestException);
-    });
-
-    it('Should throw an error if Files is not a pdf File', () => {
-      const crpFile = {
-        originalname: 'crpFile.png',
-        buffer: Buffer.from(''),
-      } as Express.Multer.File;
-      expect(() => controller.validateUploadedFile(crpFile, ['pdf'])).toThrow(
-        BadRequestException
-      );
-    });
-
-    it('Should throw an error if Files is missing', () => {
-      const crpFile = undefined;
-      expect(() => controller.validateUploadedFile(crpFile, ['pdf'])).toThrow(
-        BadRequestException
-      );
-    });
-
-    it('Should pass Validation With Correct File extention', () => {
-      const crpFile = {
-        originalname: 'crpFile.pdf',
-        buffer: Buffer.from(''),
-      } as Express.Multer.File;
-      expect(() =>
-        controller.validateUploadedFile(crpFile, ['pdf'])
-      ).not.toThrow();
-    });
-  });
-
-  describe('create', () => {
-    it('should create a psychologist with the provided files and data', async () => {
-      const files = [
-        { fieldname: 'profilePicture', originalname: 'profile.jpg' },
-        { fieldname: 'crpFile', originalname: 'crp.pdf' },
-        { fieldname: 'identifyfile', originalname: 'id.pdf' },
-        { fieldname: 'degreeFile', originalname: 'degree.pdf' },
-      ] as Express.Multer.File[];
-
-      const cpf = {
-        cpf: '123.456.789-00',
-      } as Cpf;
-
-      const psychologistData = {
-        person: {
-          rg: '98.747.153-7',
-          birthdate: new Date('2000-01-01'),
-          name: 'name de teste',
-          cpf,
-          phone: new Phone({ ddi: '+1', ddd: '123', number: '4567890' }),
-          address: {
-            zipCode: '44444-44',
-            state: 'BA',
-            street: 'Rua de Address teste',
-            city: 'cidade',
-            district: 'District de Address teste',
-            homeNumber: '10',
-            complement: 'Complement de Address teste',
-          },
-        },
-        user: {
-          email: 'john@example.com',
-          password: 'Password@123',
-          role: 'PSI',
-        },
-        crp: { crp: '00/123456' },
-        meetValue: 100,
-        meetDuration: 60,
-      };
-
-      const psychologistDataString = JSON.stringify(psychologistData);
-
-      await controller.create(files, psychologistDataString);
-
-      expect(service.create).toHaveBeenCalledWith(
-        expect.any(CreatePsychologistDto),
-        files[0], // profilePicture
-        files[1], // crpFile
-        files[2], // identifyfile
-        files[3] // degreeFile
-      );
-    });
-  });
+  
   describe('findAll', () => {
     it('Should list all Psychologists', async () => {
       expect(await controller.findAll()).toEqual([
